% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getKappa}
\alias{getKappa}
\title{Function to calculate kappa score between two terms}
\arguments{
\item{t1_genes, }{vector containing term1's geneIDs}

\item{t2_genes, }{vector containing term2's geneIDs}

\item{totalGeneCount, }{double representing count of all unique genes in dataset of interest}
}
\value{
kappaScore, double
}
\description{
Function to calculate kappa score between two terms
}
