% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KappaSimilarityMatrix}
\alias{KappaSimilarityMatrix}
\title{Create a Kappa similarity matrix}
\usage{
KappaSimilarityMatrix(myTerms, myGenes, myPvalues, kappaCutoff = 0.5)
}
\arguments{
\item{myTerms, }{A character vector (R) containing biological terms}

\item{myGenes, }{A character vector (R) where each item is a string of comma-separated geneIDs corresponding to a single term}

\item{myPvalues, }{A numeric vector (R) containing p-values corresponding to each term.}

\item{kappaCutoff, }{Numeric value specifying the cutoff threshold for kappa scores. Default is 0.5.}
}
\value{
A list containing data frames for significant term indices, all kappa scores, and the kappa score matrix.
}
\description{
This function computes kappa scores between pairs of biological terms
based on shared gene IDs and creates data frames for analysis in R.
}
