% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{_PACKAGE}
\name{RichStudio}
\alias{RichStudio}
\title{RichStudio: A package for FGSEA and clustering}
\description{
RichStudio: A package for FGSEA and clustering
}
\section{RichStudio functions}{

The RichStudio functions ...
}

